refactor: split disk_cache
